requirements.txt
Flask
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
